+++ backend/mill/minimumViableProduct/src/errorHandling/A_ValueClassesWithErrorHandling.scala
@@ -1,21 +1,24 @@
 package errorHandling
 import mvp.common.*
 
-/** =Opaque Types with Error Handling=
+/** =Value Classes with Error Handling in Scala=
   *
-  * Opaque types are a Scala 3 feature that provides type abstraction without runtime overhead. It allows to create new
-  * types that are light-weighted and can incorporate benefits of Value Classes and other structures.
+  * Value Classes provide a way to create type-safe wrappers around primitive types while maintaining runtime
+  * efficiency. When combined with companion objects and error handling, they offer a robust way to validate data at
+  * creation time. They are particularly useful in domains where data validation is crucial. This pattern helps catch
+  * errors early in the development cycle and provides clear feedback about validation failures, making systems more
+  * maintainable and reliable.
   *
-  * Basic syntax:
+  * Basic Syntax:
   * {{{
-  * opaque type MyOpaqueType = UnderlyingType
-  * object MyOpaqueType:
-  *   def apply(underlyingValue: UnderlyingType): MyOpaqueType = underlyingValue
-  *   def parse(input: UnderlyingType): Either[String, MyOpaqueType] =
+  * class ValueClass private (val value: Type) extends AnyVal
+  *
+  * object ValueClass:
+  *   def apply(value: Type): Either[Error, ValueClassType] =
   *     Either.cond(
-  *       // boolean_condition
-  *       MyOpaqueType(input),
-  *       "Error message"
+  *       boolean_condition,
+  *       new ValueClass(value),
+  *       Error("Error message")
   *     )
   * }}}
   *
@@ -24,71 +27,70 @@ import mvp.common.*
   *   - Validation control
   *   - Use of Either for error handling
   *
-  * ==Pros of Opaque Types with Error Handling==
-  *   - Performance Benefits: No runtime overhead. Opaque types do not exist during runtime.
+  * ==Pros of Value Classes with Error Handling==
+  *   - Performance Benefits: No runtime overhead at instantiation; avoids boxing/unboxing in most cases; memory
+  *     efficient compared to regular classes
   *   - Safety Guarantees: Compile-time type safety; runtime validation guarantees; immutable by design
   *   - Developer Experience: Clear API boundaries; self-documenting code; easy to maintain and refactor
-  *   - Encapsulation: Prevent invalid states through controlled construction
   *
   * ==Cons of Value Classes with Error Handling==
-  *   - Implementation Complexity: Opaque type's underlying representation is only visible in the companion object; can
-  *     make debugging more challenging
+  *   - Implementation Complexity: Requires more initial setup code
   *   - Usage Restrictions: Cannot extend other classes; Limited to a single parameter; Some scenarios force boxing
   *   - Learning Curve: Requires understanding of Either type; Pattern matching knowledge needed; (Basic) Functional
   *     programming concepts required
-  *   - Potential Overuse: Can lead to unnecessary abstraction if not used judiciously; might complicate simple code if
-  *     used where not needed
   */
 
-object E_OpaqueTypesWithErrorHandling:
+object A_ValueClassesWithErrorHandling:
 
-  private[errorHandling] opaque type Number = String
-  private[errorHandling] object Number:
+  private final class Number(val value: String) extends AnyVal
+  private object Number:
     def apply(number: String): Number =
       requireValidNumber(number)
-      number
+      new Number(number)
 
     def either(number: String): Either[InvalidNumber, Number] =
       Either.cond(
         number.forall(_.isDigit),
-        Number(number),
+        new Number(number),
         InvalidNumber(number)
       )
 
-  private[errorHandling] opaque type NieNumber = String
-  private[errorHandling] object NieNumber:
-    def apply(nieNumber: String): Number =
-      requireValidNieNumber(nieNumber)
-      nieNumber
+  private final class NieNumber(val value: String) extends AnyVal
+  private object NieNumber:
+    def apply(nieNumber: String): NieNumber =
+      val number = Number(nieNumber)
+      requireValidNieNumber(number.value)
+      new NieNumber(number.value)
 
     def either(nieNumber: String): Either[FailedValidation, NieNumber] =
       Number
         .either(nieNumber)
         .flatMap: number =>
           Either.cond(
-            number.toString.length == 7,
-            NieNumber(number),
-            InvalidNieNumber(number)
+            number.value.length == 7,
+            new NieNumber(number.value),
+            InvalidNieNumber(number.value)
           )
 
-  private[errorHandling] opaque type DniNumber = String
-  private[errorHandling] object DniNumber:
-    def apply(dniNumber: String): Number =
-      requireValidDniNumber(dniNumber)
-      dniNumber
+  private final class DniNumber(val value: String) extends AnyVal
+  private object DniNumber:
+    def apply(dniNumber: String): DniNumber =
+      val number = Number(dniNumber)
+      requireValidDniNumber(number.value)
+      new DniNumber(number.value)
 
     def either(dniNumber: String): Either[FailedValidation, DniNumber] =
       Number
         .either(dniNumber)
         .flatMap: number =>
           Either.cond(
-            number.toString.length == 8,
-            DniNumber(number),
-            InvalidDniNumber(number)
+            number.value.length == 8,
+            new DniNumber(number.value),
+            InvalidDniNumber(number.value)
           )
 
   private[errorHandling] final class DNI private (dniNumber: DniNumber, letter: ControlLetter) extends ID:
-    override def formatted: String = s"$dniNumber-$letter"
+    override def formatted: String = s"${dniNumber.value}-$letter"
 
   private[errorHandling] object DNI:
     def apply(input: String): DNI =
@@ -97,7 +99,7 @@ object E_OpaqueTypesWithErrorHandling:
       val letter = input.last.toString
       requireValidControlLetter(letter)
       val controlLetter = ControlLetter.valueOf(letter)
-      requireValidDni(dniNumber, controlLetter)
+      requireValidDni(dniNumber.value, controlLetter)
       new DNI(dniNumber, controlLetter)
 
     def either(input: String): Either[FailedValidation, DNI] =
@@ -107,15 +109,15 @@ object E_OpaqueTypesWithErrorHandling:
         dniNumber <- DniNumber.either(number)
         controlLetter <- ControlLetter.either(letter)
         result <- Either.cond(
-          dniNumber.toInt % 23 == controlLetter.ordinal,
+          dniNumber.value.toInt % 23 == controlLetter.ordinal,
           new DNI(dniNumber, controlLetter),
-          InvalidDni(dniNumber, controlLetter)
+          InvalidDni(dniNumber.value, controlLetter)
         )
       yield result
 
   private[errorHandling] final class NIE private (nieLetter: NieLetter, nieNumber: NieNumber, letter: ControlLetter)
       extends ID:
-    override def formatted: String = s"$nieLetter-$nieNumber-$letter"
+    override def formatted: String = s"$nieLetter-${nieNumber.value}-$letter"
 
   private[errorHandling] object NIE:
     def apply(input: String): NIE =
@@ -127,7 +129,7 @@ object E_OpaqueTypesWithErrorHandling:
       val letter = input.last.toString
       requireValidControlLetter(letter)
       val controlLetter = ControlLetter.valueOf(letter)
-      requireValidNie(_nieLetter, nieNumber, controlLetter)
+      requireValidNie(_nieLetter, nieNumber.value, controlLetter)
       new NIE(_nieLetter, nieNumber, controlLetter)
 
     def either(input: String): Either[FailedValidation, NIE] =
@@ -139,9 +141,9 @@ object E_OpaqueTypesWithErrorHandling:
         nieNumber <- NieNumber.either(number)
         controlLetter <- ControlLetter.either(letter)
         result <- Either.cond(
-          ((_nieLetter.ordinal * 10000000) + nieNumber.toInt) % 23 == controlLetter.ordinal,
+          ((_nieLetter.ordinal * 10000000) + nieNumber.value.toInt) % 23 == controlLetter.ordinal,
           new NIE(_nieLetter, nieNumber, controlLetter),
-          InvalidNie(_nieLetter, nieNumber, controlLetter)
+          InvalidNie(_nieLetter, nieNumber.value.toString, controlLetter)
         )
       yield result
 
@@ -169,9 +171,11 @@ object E_OpaqueTypesWithErrorHandling:
         input.trim // Handeling empty spaces around
           .replace("-", "") // Removing dashes
           .toUpperCase() // Handling lower case
+
       if !(_input.length == 9 && _input.forall(_.isLetterOrDigit))
       then Left(InvalidInput(input))
       else
+
         // Selecting which type of ID base on initial character type - Letter or Digit
         if _input.head.isDigit // Splitting between DNI and NIE
         then DNI.either(_input)
